unsafesql: refactor, move tests from authorization to unsafesql
We've added the unsafesql package recently to guard against unsafe
access to database internals. There were originally some tests in the
pkg/sql package, but we've rehomed them to unsafesql where they're more
appropriate.

Separately, we moved from using the connection pool directly provided by
the test server, to individual connections to protect the continuity of
the sessions.

Fixes: none
Epic: none
Release note: none
